Move annotate_flex_attention_for_regional_inductor to attention.py
